Moves ResistanceSet prototype to correct folder
