+++ Content.Server/GameObjects/Components/Damage/DamageableComponent.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Content.Server.Interfaces;
 using Content.Server.Interfaces.GameObjects;
 using Content.Shared.GameObjects;
+using Content.Shared.Prototypes;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Interfaces.GameObjects;
 using Robust.Shared.IoC;
